# swagger.yaml
openapi: 3.0.1
info:
  title: mlpanel
  description: ""
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
security:
  - bearerAuth: []
servers:
  - url: http://localhost:8089/
  - url: https://localhost:8089/
paths:
...
  /deployments:
    get:
      tags:
        - deployments
      summary: List deployments
      parameters:
        - name: _sort
          in: query
          description: Sort key
          required: false
          schema:
            type: string
        - name: _order
          in: query
          description: Sorting order
          required: false
          schema:
            type: string
            default: ASC
            enum:
              - ASC
              - DESC
        - name: _start
          in: query
          description: Start index
          required: false
          schema:
            type: integer
        - name: _end
          in: query
          description: End index
          required: false
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Deployment'
    post:
      tags:
        - deployments
      summary: Create deployments
      parameters:
        - name: project_id
          in: query
          description: Project ID
          required: true
          schema:
            type: integer
        - name: model_id
          in: query
          description: Model ID
          required: true
          schema:
            type: string
        - name: version
          in: query
          description: Model version
          required: true
          schema:
            type: string
            default: '1'
        - name: type
          in: query
          description: Deployment type
          required: true
          schema:
            type: string
            enum:
              - local
              - gcp
      responses:
        202:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  deployment_id:
                    description: Deployment ID
                    type: string
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
  /deployments/{deployment_id}:
    get:
      tags:
        - deployments
      summary: Get deployment
      parameters:
        - name: deployment_id
          in: path
          description: Deployment ID
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - deployments
      summary: Delete deployment (mark as deleted)
      parameters:
        - name: deployment_id
          in: path
          description: Deployment ID
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  deployment_id:
                    type: string
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
  /deployments/{deployment_id}/run:
    put:
      tags:
        - deployments
      summary: Run deployment
      parameters:
        - name: deployment_id
          in: path
          description: Deployment ID
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
               type: object
               properties:
                 deployment_id:
                   type: string
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
  /deployments/{deployment_id}/stop:
    put:
      tags:
        - deployments
      summary: Stop deployment
      parameters:
        - name: deployment_id
          in: path
          description: Deployment ID
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  deployment_id:
                    type: string
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
  /deployments/{deployment_id}/predict:
    post:
      tags:
        - deployments
      summary: Predict
      parameters:
        - name: deployment_id
          in: path
          description: Deployment ID
          required: true
          schema:
            type: integer
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - data
              properties:
                data:
                  type: object
                  properties:
                    schema:
                      type: object
                      properties:
                        fields:
                          type: array
                          items:
                            description: Column names and types
                            type: object
                            properties:
                              name:
                                description: Column name
                                type: string
                              type:
                                description: Column type
                                type: string
                    data:
                      type: array
                      items:
                        type: object
            example:
              {
                'data':
                  {
                    'schema': {
                      'fields': [
                      {'name': 'sepal_length', 'type': 'number'},
                      {'name': 'sepal_width', 'type': 'number'},
                      {'name': 'petal_length', 'type': 'number'},
                      {'name': 'petal_width', 'type': 'number'}
                      ]
                    },
                    'data': [
                    {
                      'sepal_length': 6.1,
                      'sepal_width': 2.8,
                      'petal_length': 4.7,
                      'petal_width': 1.2
                    }
                    ]
                  }
              }

        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  prediction:
                    description: Prediction result
                    type: string
        400:
          description: Bad schema
          content:
            application/json:
              schema:
                type: object
                properties:
                  anomalies:
                    $ref: '#/components/schemas/DataAnomalies'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
...
  /deployments/{deployment_id}/schema:
    get:
      tags:
        - deployments
      summary: Get deployment schema
      parameters:
        - name: deployment_id
          in: path
          description: Deployment ID
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Tensorflow data validation statistics converted to dictionary
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
  /deployments/{deployment_id}/validation-report:
    get:
      tags:
        - deployments
      summary: Get validation reports for specified period
      parameters:
        - name: deployment_id
          in: path
          description: Deployment ID
          required: true
          schema:
            type: integer
        - name: timestamp_from
          in: query
          description: Timestamp of period start
          required: true
          schema:
            type: number
            format: float
        - name: timestamp_to
          in: query
          description: Timestamp of period end
          required: true
          schema:
            type: number
            format: float
      responses:
        200:
          description: Validation report
          content:
            application/json:
              schema:
                type: object
                properties:
                  timestamp_from:
                    description: Timestamp of period start
                    type: number
                    format: float
                  timestamp_to:
                    description: Timestamp of period end
                    type: number
                    format: float
                  model_statistics:
                    description: TFDV statistics for model train data
                    type: object
                  incoming_data_statistics:
                    description: TFDV statistics for data incoming to predict for specified period
                    type: object
                  anomalies_detected:
                    description: Anomalies detected flag
                    type: boolean
                  anomalies_info:
                    $ref: '#/components/schemas/DataAnomalies'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Deployment:
      type: object
      properties:
        id:
          description: Deployment ID
          type: string
        project_id:
          description: Project ID
          type: integer
        model_id:
          description: Model ID
          type: string
        version:
          description: Model version
          type: string
        model_uri:
          description: Model URI
          type: string
        type:
          description: Deployment type
          type: string
          enum:
            - local
            - gcp
        created_at:
          description: Creation time
          type: string
          format: date-time
        host:
          description: Host address
          type: string
        port:
          description: Port number
          type: string
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
    DataAnomalies:
      type: object
      properties:
        <column_name>:
          type: object
          properties:
            description:
              description: Anomaly description
              type: string
            severity:
              description: Anomaly severity
              type: string
            shortDescription:
              description: Anomaly short description
              type: string
            reason:
              type: array
              items:
                type: object
                properties:
                  type:
                    description: Error type
                    type: string
                  shortDescription:
                    description: Anomaly short description
                    type: string
                  description:
                    description: Anomaly description
                    type: string
            path:
              type: object
              properties:
                step:
                  type: string
                  example: <column_name>
...